# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "테스트",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/posts/{count}": {
      "get": {
        "tags": ["detail"],
        "description": "",
        "parameters": [
          {
            "name": "count",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
  }
}